Updated assignTA and removeTA methods. Still troubleshooting removeTA.

+++ TAScheduler/classes/Courses.py
@@ -41,15 +41,22 @@ class CoursesClass():
     def assignTA(self, CourseCode, TAcode):
         courseCode = Course.objects.get(courseCode=CourseCode)
         taCode = MyUser.objects.get(IDNumber=TAcode)
-        # toUpdate.TAcode = MyUser.objects.get(IDNumber=TAcode)
-        #
-        temp = ClassTAAssignments(AssignmentsID='5', courseCode=courseCode, TAcode=taCode)
-        temp.save()
+        if(ClassTAAssignments.objects.get(courseCode=courseCode).TAcode == taCode):
+            raise RuntimeError("That TA is already assigned to that course!")
+        else:
+            temp = ClassTAAssignments(courseCode=courseCode, TAcode=taCode)
+            temp.save()
 
     def removeTA(self, CourseCode, TAcode):
         courseCode = Course.objects.get(courseCode=CourseCode)
-        taCode = MyUser.objects.get(IDNumber=TAcode)
-        toUpdate = ClassTAAssignments.objects.get(courseCode=courseCode)
-        toUpdate.delete()
-
-
+        if ClassTAAssignments.objects.get_queryset() == ClassTAAssignments.objects.none():
+            raise RuntimeError("No ClassTAAssignment Object exists")
+        #taCode = MyUser.objects.get(IDNumber=TAcode)
+        else:
+            toUpdate = ClassTAAssignments.objects.get(courseCode=courseCode)
+        # except RuntimeError:
+        #     print("No ClassTAAssignment Object exists")
+        if toUpdate.TAcode.IDNumber != TAcode:
+            raise RuntimeError("That TA isn't assigned to that course!")
+        else:
+            toUpdate.delete()
